Activity packet stabilizes
The issue was that the activation increases to large values, where the
difference between corresponding firing rates is not discernible. So,
the firing rate gaussian would flatten out.

To handle this, I tinkered with the sigmoid threshold in the equation
used to generate the firing rate. I increased the threshold to 1.2 times
the activation, causing the firing rate so stretch out into the
negative (negative here means < 0.5). This keeps the firing rate
gaussian well stretched and also keeps the activation in control.

Do not tinker with this value unless you must. A proper analysis of the
equations is a must to understand the stability condition in terms of
our constants.