Intermediate
I've done all the menial work. I need to work on the calibrate and
actual running methods now. I should be able to do this over the weekend
and finish it off.

Setting preferred locations might need some thinking, especially since I
need it to be structured enough to be able to generate graphs that look
2D.