Training looks good.
I can't decide if I need a backward velocity cell too. For example, what
if the new head direction is right opposite of the initial direction.
Shouldn't that make the packet move backwards?

Basically, I need a "South" training session too, that will simply go
from the end to the start.

At the moment, I have forward, left, right, but no back.

What I'm confused about is, that if the agent does turn around, won't
that be counted as going forward in the negative direction, but still
*forward*? The graph is slightly slanted towards the ends. I'm unsure
what's caused that. I still need to take a look at the P_HD_Vel
synapses. They should be fine. The logic is quite sound.